Added controller tests

+++ domain/price.go
@@ -10,10 +10,10 @@ type Price struct {
 	Market 		*Market
 	ItemId		uint
 	Item        *Item
-	Price 		int
+	Price 		float64
 	Date 		shared.DateTime
 }
 
-func NewPrice(id uint, market *Market, item *Item, price int, date shared.DateTime, ItemId uint, MarketId uint) *Price {
+func NewPrice(id uint, market *Market, item *Item, price float64, date shared.DateTime, ItemId uint, MarketId uint) *Price {
 	return &Price{Id: id, Market: market, Item: item, Price: price, Date: date, MarketId: MarketId, ItemId: ItemId}
 }
